validation

+++ app/Http/Controllers/StolenBicycleController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\StolenBicycle;
+use App\Http\Requests\StolenBicycleRequest;
 use Cloudinary;
 
 class StolenBicycleController extends Controller
@@ -16,13 +17,16 @@ class StolenBicycleController extends Controller
     public function index(Request $request, StolenBicycle $slnbike)
     {
         $hasFilter = $request->filled('model') || $request->filled('manufacturer') ||
-                     $request->filled('frame_num') || $request->filled('bouhan_num');
+            $request->filled('frame_num') || $request->filled('bouhan_num');
 
         $models = config("bicycle.models");
 
         if ($hasFilter) {
             $slnbikes = $slnbike->getFiltered($request->only([
-                'model', 'manufacturer', 'frame_num', 'bouhan_num'
+                'model',
+                'manufacturer',
+                'frame_num',
+                'bouhan_num'
             ]));
         } else {
             $slnbikes = $slnbike->getPaginateByLimit();
@@ -45,7 +49,7 @@ class StolenBicycleController extends Controller
         return view('stolen_bicycles.create')->with(['models' => $models, 'manufacturers' => $manufacturers, 'prefectures' => $prefectures,]);
     }
 
-    public function store(Request $request)
+    public function store(StolenBicycleRequest $request)
     {
         $slnbike = new StolenBicycle();
         if ($request->hasFile('image')) {
@@ -54,47 +58,40 @@ class StolenBicycleController extends Controller
         $slnbike->user_id = auth()->id();
         $input = $request['stolenbicycle'];
         $slnbike->fill($input)->save();
-        return redirect('/stolenbicycles/' . $slnbike->id);
+        return redirect()->route('sln.show', $slnbike);
     }
 
     public function delete(StolenBicycle $slnbike)
     {
-        // if (!auth()->check()) {
-        //     // 未ログインならログインページへ
-        //     return redirect()->route('login'); // または abort(401)
-        // }
-
-        // if (auth()->id() !== $slnbike->user_id) {
-        //     abort(403, '許可されていません');
-        // }
+        if (auth()->id() !== $slnbike->user_id) {
+            abort(403, '権限がありません');
+        }
         $slnbike->delete();
-        return redirect('/stolenbicycles');
+        return redirect()->route('sln.show');
     }
 
     public function edit(StolenBicycle $slnbike)
     {
-        // if (!auth()->check()) {
-        //     // 未ログインならログインページへ
-        //     return redirect()->route('login'); // または abort(401)
-        // }
-
-        // if (auth()->id() !== $slnbike->user_id) {
-        //     abort(403, '権限がありません．');
-        // }
+        if (auth()->id() !== $slnbike->user_id) {
+            abort(403, '権限がありません');
+        }
         $models = config("bicycle.models");
         $manufacturers = collect(config("bicycle.manufacturers"))->sortKeys();
         $prefectures = config("bicycle.prefectures");
         return view('stolen_bicycles.edit')->with(['slnbike' => $slnbike, 'models' => $models, 'manufacturers' => $manufacturers, 'prefectures' => $prefectures,]);
     }
 
-    public function update(Request $request, StolenBicycle $slnbike)
+    public function update(StolenBicycleRequest $request, StolenBicycle $slnbike)
     {
+        if (auth()->id() !== $slnbike->user_id) {
+            abort(403, '権限がありません');
+        }
         if ($request->hasFile('image')) {
             $slnbike->image_path = Cloudinary::upload($request->file('image')->getRealPath(), ['public_id' => 'slnbike_' . $slnbike->id, 'overwrite' => true])->getSecurePath();
         }
         $slnbike->user_id = auth()->id();
         $input = $request['stolenbicycle'];
         $slnbike->fill($input)->save();
-        return redirect('/stolenbicycles/' . $slnbike->id);
+        return redirect()->route('sln.show', $slnbike);
     }
 }
